fix: Address 16 MAJOR SonarCloud issues (S1066, S1172)
Changes:
- Merge 8 nested if statements using short-circuit && (S1066)
  config.c, parser_utils.c, main.c, lrc_parser.c, lyrics_provider.c, system_tray.c
- Suppress 8 unused parameter warnings with (void) casts in GDBus callbacks (S1172)
  dbus_control.c: handle_method_call, on_bus_acquired, on_name_acquired, on_name_lost

+++ src/user_experience/config/config.c
@@ -1249,12 +1249,11 @@ static char* try_load_user_config(struct config *cfg, const char *path) {
     umask(old_mask);
 
     // If file was just created (fd >= 0), copy from system config
-    if (fd >= 0) {
-        if (!copy_system_config_to_user(fd, path)) {
-            // Copy failed (system config doesn't exist) - remove empty file
-            unlink(path);
-            return NULL;
-        }
+    if (fd >= 0 &&
+        !copy_system_config_to_user(fd, path)) {
+        // Copy failed (system config doesn't exist) - remove empty file
+        unlink(path);
+        return NULL;
     }
     // If fd < 0 and errno == EEXIST, file already exists - continue normally
